Fix generate-types
The shapes reference was pointing to share which was outside of the root directory and causing problems.
