+++ how-to/customize-workspace/client/src/framework/shapes/index.ts
@@ -14,3 +14,4 @@ export * from "./store-shapes";
 export * from "./bootstrap-shapes";
 export * from "./browser-shapes";
 export * from "./app-shapes";
+export * from "./share-shapes";
