+++ how-to/customize-workspace/client/src/framework/shapes/share-shapes.ts
@@ -1,0 +1,25 @@
+import type { Page } from "@openfin/workspace-platform";
+
+export interface IShareCustomData {
+	workspaceId?: string;
+	windowIdentity?: OpenFin.Identity;
+	pageId?: string;
+	page?: Page;
+	bounds?: OpenFin.Bounds;
+}
+
+export type IShareEntry = IShareEntryPage | IShareEntryWorkspace | IShareEntryUnknown;
+
+interface IShareEntryPage {
+	type: "page";
+	data: { page: Page; bounds: OpenFin.Bounds };
+}
+
+interface IShareEntryWorkspace {
+	type: "workspace";
+	data: { snapshot: string };
+}
+
+interface IShareEntryUnknown {
+	type: "other";
+}
